bot/requirements.txt
python-telegram-bot==21.10
google-generativeai==0.8.5
openai==1.82.0
groq==0.26.0
gTTS==2.5.4
httpx==0.28.1
